backend/requirements.txt
flask
flask-cors
pyngrok
supabase
python-dotenv
google-genai
